Updated deviations and CSW
CSW still needs work, Deviations needs more testing

+++ DataInterface/IBcsoRepository.cs
@@ -4,6 +4,6 @@ namespace DataInterface
 {
     public interface IBcsoRepository : IRepository<Bcso>
     {
-        double GetAmount(double income, int numberOfChildren);
+        double GetAmount(int income, int numberOfChildren);
     }
 }
