+++ Models/Bcso.cs
@@ -5,7 +5,7 @@ namespace Models
     public class Bcso : IEntity
     {
         public long Id { get; set; }
-        public double GrossIncome { get; set; }
+        public int GrossIncome { get; set; }
         public double OneChildAmount { get; set; }
         public double TwoChildAmount { get; set; }
         public double ThreeChildAmount { get; set; }
